TO: fix
git-svn-id: svn://svn.code.sf.net/p/boomerang/code/trunk@2431 dc5e1367-85d4-4bd1-96b4-a20bfbfd92d5

+++ boomerang/frontend/pentiumdecoder.h
@@ -37,13 +37,13 @@ public:
 	 * Decodes the machine instruction at pc and returns an RTL instance for
 	 * the instruction.
 	 */
-virtual DecodeResult& decodeInstruction (ADDRESS pc, int delta);
+virtual DecodeResult& decodeInstruction (ADDRESS pc, intptr_t delta);
 
 	/*
 	 * Disassembles the machine instruction at pc and returns the number of
 	 * bytes disassembled. Assembler output goes to global _assembly
 	 */
-virtual int decodeAssemblyInstruction (ADDRESS pc, int delta);
+virtual int decodeAssemblyInstruction (ADDRESS pc, intptr_t delta);
 
 private:
 	/*
@@ -61,7 +61,7 @@ private:
 	SWord	getWord(unsigned lc);
 	DWord	getDword(unsigned lc);
 
-	unsigned lastDwordLc;
+	uintptr_t lastDwordLc;
 };
 
 #endif
